v2/requirements.txt
alpaca-py>=0.21.0
anthropic>=0.40.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2025.2
flask>=3.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
